Add support for wildcard $select parameter
Limiting the properties returned for an entity by using the $select
parameter worked while explicit property names were passed, however
the odata spec also allows wildcard properties.  This patch adds that
support.  So now, a request such as Products(1)?$select=* will return
all properties of the Products entity.

Using the wildcard through navigation properties is not supported by
this patch as the expression of navigation properties in $select is
not supported.  See issue 249.

Bug: Issue 163
Change-Id: Ia318c41b567ca2db76949e826ab6580323f65542