Add RestEasy support, Visitor framework changes
Using the RestEasy as the REST engine with odata4j does not work as
direct context injection of ContextResolver used by odata4j is not
supported in RestEasy. After reading JSR-311, found that the context
injection style used in odata4j is not defined by the spec thus
not supported by the RestEasy.  This fix addresses the issue by
injecting @context Providers, and then reading the ContextResolvers
from it. This solution works with all the JAX-RS engines (cxf,jersy,
RestEasy)

Please note that to fully work with RestEasy one also needs to be
working with release of RestEasy that has
https://issues.jboss.org/browse/RESTEASY-812

The current visitor framework is restrictive, as it dictates one way
of the language object traversal, adding additional visitors with
different language object traversal behavior is difficult. This fix
addresses the issue, by making visiting and hierarchy traversal
as orthogonal issues, and provides a more flexible and
extensible framework.

OData specification supports fully qualified entity names, support
for fully qualified names not seems to be fully supported by odata4j
in all the code paths. This fix addresses the issue.

Bug: Issue 238
Bug: Issue 239
Bug: Issue 254

Change-Id: Ia8b0e0da91e2e7c6caaffa9af39b53d43db86c4f